rebase and fix conflicts
