+++ sdk/keyvault/perf-tests/keyvault-secrets/test/index.spec.ts
@@ -1,5 +1,6 @@
 import { PerfStressProgram, selectPerfStressTest } from "@azure/test-utils-perfstress";
-import { GetSecretTest, ListSecretsTest } from "./getSecret.spec";
+import { GetSecretTest } from "./getSecret.spec";
+import { ListSecretsTest } from "./listSecrets.spec";
 
 console.log("=== Starting the perfStress test ===");
